+++ test/templates/landingpage/content-elements/content/column-2/index.js
@@ -1,23 +1,28 @@
-const {cx, Icon} = require('@bsi-cx/design-build');
-const {CX_23_2} = require('../../../../../../src/version');
+require('./styles.scss');
 
-const element = cx.contentElement;
+const { cx, Icon } = require("@bsi-cx/design-build");
 
-module.exports = element;
-
-element.withElementId('column-2')
+module.exports = cx.templateElement
+  .withElementId("column-2")
   .withIcon(Icon.TWO_COLUMNS)
-  .withLabel('2 Columns')
-  .withFile(require('./template.twig'))
-  .withArchivedMinVersion(CX_23_2)
+  .withLabel("2 Columns")
+  .withFile(require("./template.hbs.twig"))
   .withDropzones(
+    cx
+      .Dropzone("20816df1-f8c0-47d1-94a1-1cd124c2b348", [
+        require("../column-1"),
+        require("../title"),
+      ]),
     cx.dropzone
-      .withDropzone('20816df1-f8c0-47d1-94a1-1cd124c2b348')
+      .withDropzone("5971732b-bf41-493d-a678-0fce1a2b5771")
       .withAllowedElements(
-        require('../column-1'),
-        require('../title')),
-    cx.dropzone
-      .withDropzone('5971732b-bf41-493d-a678-0fce1a2b5771')
-      .withAllowedElements(
-        require('../column-1'),
-        require('../title')));
+        require("../column-1"),
+        require("../title"),
+        require("../text"),
+      ),
+  )
+  .withScopePrefills(
+    cx
+      .ScopePrefill("buttonLeft", require("../../../../website/content-elements/content/template-button"))
+      .withOverrideValue("multiline-plain-text-wmiRti", "Andere Info"),
+  );