+++ test/templates/website/content-elements/content/column-2/index.js
@@ -10,12 +10,7 @@ module.exports = cx.templateElement
       .Dropzone("20816df1-f8c0-47d1-94a1-1cd124c2b348", [
         require("../column-1"),
         require("../title"),
-      ])
-      .withScopePrefills(
-        cx
-          .ScopePrefill("buttonLeft", require("../template-button"))
-          .withOverrideValue("multiline-plain-text-wmiRti", "Andere Info"),
-      ),
+      ]),
     cx.dropzone
       .withDropzone("5971732b-bf41-493d-a678-0fce1a2b5771")
       .withAllowedElements(
@@ -24,4 +19,9 @@ module.exports = cx.templateElement
         require("../text"),
         require("../image"),
       ),
+  )
+  .withScopePrefills(
+    cx
+      .ScopePrefill("buttonLeft", require("../template-button"))
+      .withOverrideValue("multiline-plain-text-wmiRti", "Andere Info"),
   );